7b: move GrabFrameFileUtilities and FileAssociationUtilities to Core.Windows
Section 7 had both files marked "unblocked by a8591aa, never audited past
that blocker" - HistoryInfo (their root pin to the app) moved to Core.Windows
in that commit, but nobody had re-read either file since to check for a
second blocker. This batch is that audit, and the answer is: no second
blocker, both move clean.

GrabFrameFileUtilities.cs needed zero code changes - it already traffics
entirely in System.Drawing.Bitmap (Core.Windows-legal per B3),
System.IO/System.IO.Compression, System.Text.Json, and HistoryInfo/
WordBorderInfo, both already in Core.Windows. Its only non-BCL call
(AutomationProfile.GetTemporaryDirectory()) has been internal-and-Core since
6a, and Core.Windows already has InternalsVisibleTo from Core for exactly
this kind of call.

FileAssociationUtilities.cs needed one substitution, the same one 4e/6a
established for this exact pattern: AppUtilities.IsPackaged() ->
PackageIdentity.IsPackaged(). Everything else (Microsoft.Win32.Registry,
FileUtilities.GetExePath(), GrabFrameFileUtilities.GrabFrameFileExtension)
was already Core.Windows-legal. Text-Grab.csproj's InternalsVisibleTo grant
to itself is irrelevant now that the class lives in Core.Windows, which
already grants InternalsVisibleTo("Text-Grab"), so the app's one call site
(App.xaml.cs) needed no change.

That move also retired Text-Grab/Utilities/OpenDocumentFilterUtilities.cs,
which existed for exactly one reason - GetOpenDocumentFilter() needed
GrabFrameFileUtilities, which needed to stay app-side. With that gone,
GetOpenDocumentFilter() folded back into FileUtilities.GetOpenDocumentFilter()
in Core.Windows (its two now-unused-elsewhere helpers, GetExtensionsFilterPattern
and GetVisualDocumentFilterPattern, went back to private), and the two call
sites (App.xaml.cs, EditTextWindow.xaml.cs) call FileUtilities directly.

Tests followed the production code: GrabFrameFileTests.cs (13 test cases,
all already plain [Fact]/[Theory] with no WPF dependency - its `using
System.Windows;` was unused even before this move) moved wholesale to
Tests.Core.Windows. GetOpenDocumentFilter_IncludesVisualAndTextOptions moved
from Tests/FilesIoTests.cs into Tests.Core.Windows/FileUtilitiesTests.cs
alongside the other FileUtilities filter test. Net effect on the gate: Tests
475 (was 489), Tests.Core.Windows 158 (was 144), same 14 cases moved across,
total still 1516.
