+++ Text-Grab.Core.Windows/Utilities/GrabFrameFileUtilities.cs
